+++ py-polars/src/lazy/dsl.rs
@@ -34,75 +34,6 @@ impl From<dsl::Expr> for PyExpr {
     }
 }
 
-#[pyclass]
-#[derive(Clone)]
-pub struct When {
-    predicate: PyExpr,
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct WhenThen {
-    predicate: PyExpr,
-    then: PyExpr,
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct WhenThenThen {
-    inner: dsl::WhenThenThen,
-}
-
-#[pymethods]
-impl When {
-    pub fn then(&self, expr: PyExpr) -> WhenThen {
-        WhenThen {
-            predicate: self.predicate.clone(),
-            then: expr,
-        }
-    }
-}
-
-#[pymethods]
-impl WhenThen {
-    pub fn when(&self, predicate: PyExpr) -> WhenThenThen {
-        let e = dsl::when(self.predicate.inner.clone())
-            .then(self.then.inner.clone())
-            .when(predicate.inner);
-        WhenThenThen { inner: e }
-    }
-
-    pub fn otherwise(&self, expr: PyExpr) -> PyExpr {
-        dsl::ternary_expr(
-            self.predicate.inner.clone(),
-            self.then.inner.clone(),
-            expr.inner,
-        )
-        .into()
-    }
-}
-
-#[pymethods]
-impl WhenThenThen {
-    pub fn when(&self, predicate: PyExpr) -> WhenThenThen {
-        Self {
-            inner: self.inner.clone().when(predicate.inner),
-        }
-    }
-    pub fn then(&self, expr: PyExpr) -> WhenThenThen {
-        Self {
-            inner: self.inner.clone().then(expr.inner),
-        }
-    }
-    pub fn otherwise(&self, expr: PyExpr) -> PyExpr {
-        self.inner.clone().otherwise(expr.inner).into()
-    }
-}
-
-pub fn when(predicate: PyExpr) -> When {
-    When { predicate }
-}
-
 pub fn col(name: &str) -> PyExpr {
     dsl::col(name).into()
 }
